// src/modules/order/models/responses/CustomerResponse.ts
export interface CustomerResponse {
  id: number;
  code: string;
  assignedDate: string | null;
  assignedStore: string | null;
  assignedStoreId: number | null;
  averageOrderAmount: number | null;
  birthday: number | null;
  birthdayCompare: number | null;
  birthdayDayAndMonthCompare: number | null;
  birthdayDayAndYearCompare: number | null;
  birthdayMonthAndYearCompare: number | null;
  cardNumber: string | null;
  channelIds: Array<string> | null;
  city: string | null;
  cityId: number | null;
  company: string | null;
  customerGroup: string | null;
  customerGroupId: number | null;
  customerLevel: string | null;
  customerLevelId: number | null;
  customerRelationship: string | null;
  customerType: string | null;
  customerTypeId: number | null;
  dayOfBirth: number | null;
  district: string | null;
  districtId: number | null;
  ecommerceUid: string | null;
  email: string | null;
  familyInfo: string | null;
  firstOrderTime: string | null;
  firstOrderTimeOffline: string | null;
  firstOrderTimeOnline: string | null;
  firstOrderType: string | null;
  fullAddress: string;
  fullName: string;
  gender: string;
  grossSale: string | null;
  identityNumber: string | null;
  lastOrderTime: string | null;
  lastOrderTimeOffline: string | null;
  lastOrderTimeOnline: string | null;
  lastOrderType: string | null;
  monthOfBirth: number | null;
  phone: string;
  point: number | null;
  responsibleStaff: null | string;
  responsibleStaffCode: null | string;
  sourceIdOfFirstOrderOnline: null | number;
  sourceIdOfLastOrderOnline: null | number;
  sourceIds: Array<number> | null;
  sourceOfFirstOrderOnline: null | string;
  sourceOfLastOrderOnline: null | string;
  storeIdOfFirstOrderOffline: null | Array<number>;
  storeIdOfLastOrderOffline: null | Array<number>;
  storeIds: null | Array<number>;
  storeOfFirstOrderOffline: number | null;
  storeOfLastOrderOffline: number | null;
  totalFinishedOrder: number | null;
  totalOrderHasGrossSale: number | null;
  totalPaidAmount: number | null;
  totalRefundedAmount: number | null;
  totalReturnedOrder: number | null;
  utmCampaign: string | null;
  utmContent: string | null;
  utmId: string | null;
  utmMedium: string | null;
  utmSource: string | null;
  utmTerm: string | null;
  ward: string | null;
  wardId: number | null;
  website: string | null;
  weddingDate: string | null;
  weddingDay: number | null;
  weddingMonth: number | null;
  weddingYear: number | null;
  yearOfBirth: number | null;
}